Instead of requiring calls to `ALTableViewCellFactory` from within `UITableViewDataSource` and `UITableViewDelegate` methods (on the owning view controller), `ALTableVIewCellFactory` should instead call these.
To do such, `ALTableViewCellFactory` now sets itself as the `delegate` and `datasource` for the table view passed to it via `initWithTableView:`. A few methods were also added to `ALTableViewCellFactoryDelegate`.

Why should we even want to do this? It will drastically reduce/eliminate boilerplate code related to using `ALTableViewCellFactory`. The owning view controller doesn't even need to conform to `UITableViewDataSource` or `UITableViewDelegate` if it doesn't want to! If it does, non-implemented methods by `ALTableViewCellFactory` will still be called on the view controller! See the notes on `ALTableViewCellFactoryDelegate` related to this and the updated demo project.

Note: the demo and static library are now in the same workspace.