Bumped pod spec version to `0.6.0`.
WARNING:  these version introduces changes that are NOT backwards-compatible. Until this library reaches version `1.0`, it's possible that such changes may happen from time to time.
It should be fairly easy to update an application that depends on this pod; see `ALTableViewCellFactoryDelegate` for changes in method signatures and new required methods.
If you're not ready to update your application immediately, you can simply depend on version '~> 0.5.0' in your pod file.