It seems that iOS 8 automatic height calculations have some bugs... :/ Fortunately, `ALTableViewCellFactory`, which calculates cell heights itself, still works really well. :]
In such, I've gone in a new direction on `ALTableViewManager`, which now uses `ALTableViewCellFactory`. Currently, this is very much "will this work in production?" designed code...

This really shouldn't be on master, but I also don't want to use the not workable solution from previous commits as-is on master... Honestly, I was too optimistic with committing it directly to master before...

I'm committing these last changes to master, but I'm switching to `development` for further work on this.